refactor: movement of superdroplets across domain in seperate structure to MoveSupersInDomain
