+++ libs/cartesiandomain/movement/CMakeLists.txt
@@ -3,20 +3,16 @@ if(NOT DEFINED CMAKE_MINIMUM_REQUIRED_VERSION)
   cmake_minimum_required(VERSION 3.18.0)
 endif()
 
-# set library name and print directory of this CMakeLists.txt (source directory of library) and print directory of this CMakeLists.txt (source directory of library)
-set(LIBNAME "cartesiandomain")
+# set library name and print directory of this CMakeLists.txt (source directory of library)
+set(LIBNAME "movement")
 message("${LIBNAME} LIBRARY_SOURCE_DIR: ${CMAKE_CURRENT_SOURCE_DIR}")
 
 # explicitly set library executables path to /lib in top level of build tree
-set(LIB_BINARY_DIR ${CLEO_BINARY_DIR}/lib)
-
+set(LIB_BINARY_DIR ${CMAKE_BINARY_DIR}/lib)
 
 # Add executables and create library target
 set(SOURCES
-"add_supers_at_domain_top.cpp"
-"cartesianmaps.cpp"
-"createcartesianmaps.cpp"
-"cartesian_decomposition.cpp"
+  "cartesian_transport_across_domain.cpp"
 )
 # must use STATIC (not(!) SHARED) lib for linking to executable if build is CUDA enabled with Kokkos
 add_library("${LIBNAME}" STATIC ${SOURCES})
@@ -25,8 +21,9 @@ add_library("${LIBNAME}" STATIC ${SOURCES})
 # target_include_directories(${LIBNAME} PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 target_include_directories(${LIBNAME} PRIVATE "${CLEO_SOURCE_DIR}/libs") # CLEO libs directory
 
-# Link libraries to target library
-set(LINKLIBS initialise superdrops gridboxes movement)
+# Add directories and link libraries for target library
+# target_include_directories(${LIBNAME} PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+set(LINKLIBS superdrops gridboxes)
 target_link_libraries(${LIBNAME} PUBLIC "${LINKLIBS}")
 target_link_libraries(${LIBNAME} PUBLIC Kokkos::kokkos)
 
@@ -39,5 +36,3 @@ set_target_properties(${LIBNAME} PROPERTIES
   CMAKE_CXX_STANDARD_REQUIRED ON
   CMAKE_CXX_EXTENSIONS ON
   CXX_STANDARD 20)
-
-add_subdirectory(movement)
