+++ libs/cartesiandomain/null_boundary_conditions.hpp
@@ -27,6 +27,7 @@
 #include "cartesiandomain/cartesianmaps.hpp"
 #include "gridboxes/supersindomain.hpp"
 
+// TODO(CB): make boundary conditions a concept and move into gridboxes library
 struct NullBoundaryConditions {
   SupersInDomain operator()(const CartesianMaps &gbxmaps, viewd_gbx d_gbxs,
                             const SupersInDomain &allsupers) const {
