+++ roughpaper/src/main_impl.hpp
@@ -31,6 +31,7 @@
 #include <string_view>
 
 #include "zarr/dataset.hpp"
+#include "cartesiandomain/movement/cartesian_movement.hpp"
 #include "cartesiandomain/add_supers_at_domain_top.hpp"
 #include "cartesiandomain/cartesianmaps.hpp"
 #include "cartesiandomain/cartesianmotion.hpp"
@@ -120,11 +121,10 @@ inline auto create_boundary_conditions(const Config &config) {
 
 template <GridboxMaps GbxMaps>
 inline auto create_movement(const Config &config, const Timesteps &tsteps, const GbxMaps &gbxmaps) {
-  const auto motion(create_motion(tsteps.get_motionstep()));
-
-  const auto boundary_conditions(create_boundary_conditions(config));
-
-  return MoveSupersInDomain(gbxmaps, motion, boundary_conditions);
+  const auto motion = create_motion(tsteps.get_motionstep());
+  const auto boundary_conditions = create_boundary_conditions(config);
+  const auto movement = cartestian_movement(gbxmaps, motion, boundary_conditions);
+  return movement;
 }
 
 inline MicrophysicalProcess auto config_condensation(const Config &config,
